@@ -2,18 +2,12 @@ package com.example.yhao.floatwindow;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.View;
-import android.view.animation.BounceInterpolator;
-import android.widget.ImageView;
-import android.widget.Toast;
 
 import com.example.yhao.fixedfloatwindow.R;
-import com.yhao.floatwindow.FloatWindow;
-import com.yhao.floatwindow.MoveType;
-import com.yhao.floatwindow.Screen;
 
-public class A_Activity extends AppCompatActivity {
+public class ActivityA extends AppCompatActivity {
 
 
     @Override
@@ -24,6 +18,6 @@ public class A_Activity extends AppCompatActivity {
     }
 
     public void change(View view) {
-        startActivity(new Intent(this, B_Activity.class));
+        startActivity(new Intent(this, ActivityB.class));
     }
 }
