@@ -5,7 +5,7 @@ import android.view.View;
 
 import com.example.yhao.fixedfloatwindow.R;
 
-public class C_Activity extends BaseActivity {
+public class ActivityC extends BaseActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
